Time: 99 ms (47.48%), Space: 99 MB (61.54%) - LeetHub

+++ 3332-minimum-operations-to-exceed-threshold-value-ii/3332-minimum-operations-to-exceed-threshold-value-ii.cpp
@@ -18,7 +18,7 @@ public:
             ll s2=pq.top();
             pq.pop();
 
-            long long value= (min(s1,s2) * 2) + max(s1,s2);
+            ll value= (min(s1,s2) * 2) + max(s1,s2);
             count++;
             pq.push(value);
             // cout<<value<<" ";
